# Conflicts:
#	CHANGELOG.md
#	README.md
#	oc-admin/themes/modern/css/main.css
#	oc-admin/themes/modern/tools/upgrade.php
#	oc-includes/assets/bootstrap/README.md
#	oc-includes/assets/bootstrap/bootstrap.min.js.map
#	oc-includes/osclass/classes/actions/ItemActions.php
#	oc-includes/osclass/classes/form/FieldForm.php
#	oc-includes/osclass/classes/upgrade/Osclass.php
#	oc-includes/osclass/classes/utility/Sanitize.php
#	oc-includes/osclass/default-constants.php

+++ oc-includes/osclass/helpers/hPreference.php
@@ -648,11 +648,11 @@ function osc_update_core_json()
 /**
  * Gets current version
  *
- * @return int
+ * @return string
  */
 function osc_version()
 {
-    return (int)getPreference('version');
+    return getPreference('version');
 }
